allow manual float height
i overlooked this use case before
rendering the board is always a cropping of the full board, it doesn't affect the rendering process

+++ Fumen Utils_files/rendering/modified-fumen-canvas.js
@@ -3,7 +3,7 @@ import encode64 from "../lib/b64.js"
 // import { GIFEncoder, quantize, applyPalette } from 'https://unpkg.com/gifenc@1.0.3';
 
 function getFumenMaxHeight(...fumenPages) {
-	if (!document.getElementById('autoheight').checked) return parseInt(document.getElementById('height').value)
+	if (!document.getElementById('autoheight').checked) return parseFloat(document.getElementById('height').value)
 
 	var highestRow = Math.max(...fumenPages.map(highestPageHeight))
 	return Math.max(1, Math.min(23, highestRow))
